Vehicule Edit view updated
The vehicule edit view was update to contain tiles and forms, the style for tiles and modals was added.
The routes login, logout and register, were moved to auth module, to resemble the backend routes structure.
The Pagination component was created and updates active page correctly, still the use on other modules is not finished, since it cannot execute action based on active page.