Pagination Working
Modified Pagination component and it's now a completely working component.
The vehicules get routes was modified to use new Pagination component.
A new context module was created for containing all context requeried for the project, a skeleton was added for vehicules, and is plan to be used to share modal information.