Vehicule Context for Modal
A context for vehicules get route was created for containing vehicule_id of vehicule to be deleted, in order to be user by Parent route View to display modal and execute api request.
The context needs a rename of structs.
The modal is only resetting context value on close and delete button but not on external and header button, due to being part of modal component and having hardcoded onclick callbacks