Vehicule Table Modified
The VehiculeTable and VehiculeTableRow were modified to add extra vehicule information[C, also only showing detailed information to admin.
The Modal was added an extra prop to contain an optional onclose callback to perform before closing the modal.
The VehiculeGetView was modified to add onclose prop to modal and use vehicule context to show info when clicking button.
The VehiculeItemAction was added a new variant Get for getting vehicule info to show (This option may be removed due to table showing all necessary data).