Collapsible Addeed, Forms refactor
The bulma-collapsible sass module was added to style, also a simple component for collapsible was created.
The Validation of forms was simplified by form utils and utilizing functions on each form.
The VehiculeEditView was modified and working, a VehiculeEditForm component was created and content from the view will be moved to this component.
Removed some unused imports on some modules.