Upload Component
Base64 dependency added.
A component for handling uploading of files has been created, and is able to show the file name along content if it's image or video and capable of limiting maximum file upload, deleting file. But it still does not have a way of utilizing the file data (as in API fetch) other than rendering directly as a review