Refactor of proyect
The routes module was refactor to separate between admin and normal user routes.
The Pictures and Picture components were modified to use a state and a custom wrapper for multipart that is converted into reqwest multipart Form when uploading to server.
The vehicules route were combine into a single route with modals and a reducer was created for handling the route events, but possibly might be a regression to individual routes as before