Return to individual routes instead of modal
The modal for adding vehicule was change to it's own route as before.
The edit vehicule view was split into modules on the edit module.
The edit view and vehicules (table) view were updated to get the actual image of vehicule instead of fixed dummy image.
The Vehicule type was added some useful methods for status and active translation to spanish and a method to create picture url given a base url.
The sidebar has a dropdown menu for vehicule routes as before.
The picture component now handles the multipart by a upload state handled by the parent component.
The EditVehiculeForm component upload the new picture to server given by Picture component, but doesn't reset it after successful update and does n't show loading animation when uploading