Project Structure Changed
The structure of the project was changed and the features and pages modules were created, the features mod contains all the components, services, etc to execute a feature in the system, the pages mod contains entry points to access the features, and the route mod switch between all the pages.
The components module was reduced and reorganized, keeping only basic reusable components across all features and pages, and the specific components were moved to the apropiate feature component.
The Users View was created and the basic structure was added.