Modified SCSS
A new module was added to have extend hero to hero-background, allowing to show a picture in the background of a bulma hero.
The home, login, signup pages were updated from hero to hero-background.
The SCSS table module was modified to show picture the complete picture in image-cell when viewing table in mobile.
The vehicule table was modified to use a figure html element as parent of the img.
The image on vehicule table was modified to have onclick event to open modal and show picture as the action button.
The user table was modified to show the appropiate table head and rows