Added Feature
The departments feature was added, and contains a simple DepartmentDropdown menu, for selection.
Components in users were moved into a sub folder admin, same change will be applied to vehicules.
The admin reducer in users was changed and is now working in the admin view.
A custom logo file was added to static folder and register in index.html and utilized in navbar.
A new class was created for showing images in tiles in scss.