Calendar Added, Normal pages added, Error modified
The bulma calendar component was added along it's sass.
The Normal module was added to the pages module in order to split the admin and normal user pages.
The features module was completly removed, and views moved into appropiate pages.
The profile view was added to both normal and admin, but might only be used in normal and admin is redirected to a user edit view on itself.
The error type was changed to contain a error message in the BadRequest variant if available.