Fixed components bugs, New Components, Refactor
The Calendar component was fixed to use lang weekdays on header.
The Timepicker component was created, but still unsure of what version to use, default input with type time or custom component.
The vehicules module in page/normal was moved to as a submodule of request. Vehicules sidebar entry should be removed./tuxguitar.sh
Some unused imports were removed from files.